[MOIDFY] remover.py Adding file and place functions to remove a file from Makefile
